+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Not released
 
-- Bump carto-react 1.2.1-alpha.1[#316](https://github.com/CartoDB/carto-react-template/pull/317)
+- Bump versions: carto-react 1.2.1-beta.1 and deck.gl 8.7.0-beta.2 [#320](https://github.com/CartoDB/carto-react-template/pull/320)
+- Bump carto-react 1.2.1-alpha.1 [#317](https://github.com/CartoDB/carto-react-template/pull/317)
 - Add DrawingTool in template-sample-app-3 [#314](https://github.com/CartoDB/carto-react-template/pull/314)
 - Preparing alpha 1.2.0 version of the template [#313](https://github.com/CartoDB/carto-react-template/pull/313)
 - Bump 8.7.0-alpha.11 [#316](https://github.com/CartoDB/carto-react-template/pull/316)